requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aioice==0.7.6
aiortc==1.4.0
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
##asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
Babel==2.12.1
##backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.11.2
bidict==0.22.1
##black @ file:///C:/ci/black_1660239974023/work
bleach==6.0.0
cachetools==5.3.0
##certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
##click @ file:///C:/ci/click_1646038595831/work
##cloudpickle @ file:///C:/b/abs_3796yxesic/croot/cloudpickle_1683040098851/work
##colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
##comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
ConfigArgParse==1.5.3
contourpy==1.0.7
cryptography==38.0.0
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
##decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
descartes==1.1.0
dnspython==2.3.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring-parser==0.14.1
##exceptiongroup @ file:///C:/b/abs_25wqfvkf25/croot/exceptiongroup_1668714345637/work
##executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.16.3
ffmpeg==1.4
filelock==3.9.0
fire==0.5.0
Flask==2.2.3
flatbuffers==23.5.26
##flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.5
frozenlist==1.3.3
fsspec==2023.9.2
functorch==1.13.1
gast==0.4.0
gdown==4.6.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.51.3
gviz-api==1.10.0
h5py==3.8.0
idna==3.4
imageio==2.26.0
##importlib-metadata @ file:///C:/b/abs_20ndzb2j6v/croot/importlib-metadata_1678997085534/work
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.10
##jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.19.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.0
lpips==0.1.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
##matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mdurl==0.1.2
mediapy==1.1.6
mistune==2.0.5
ml-dtypes==0.1.0
msgpack==1.0.4
msgpack-numpy==0.4.8
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.5.0
nerfacc==0.3.4
##-e git+https://github.com/nerfstudio-project/nerfstudio@b698f00237fc64e8d82c83b25c5878a30ec7ba6d#egg=nerfstudio
##nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
netifaces==0.11.0
networkx==3.0
ninja==1.11.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
nuscenes-devkit==1.1.9
oauthlib==3.2.2
open3d==0.16.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.2
pandocfilters==1.5.0
##parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
##pathspec @ file:///C:/b/abs_9cu5_2yb3i/croot/pathspec_1674681579249/work
pathtools==0.1.2
##pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pipreqs==0.4.13
##platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
plotly==5.13.1
point-cloud-utils==0.30.0
prometheus-client==0.16.0
##prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.20.3
##psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
##pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycocotools==2.0.6
pycparser==2.21
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
PyMCubes==0.1.4
pymeshlab==2022.2.post2
pyngrok==5.2.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
##python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-engineio==4.3.4
python-json-logger==2.0.7
python-socketio==5.7.2
pytz==2023.3
PyWavefront==1.3.3
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
shapely==2.0.1
shtab==1.5.8
##six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
##spyder-kernels @ file:///C:/b/abs_e788a8_4y9/croot/spyder-kernels_1691599588437/work
##stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tenacity==8.2.2
tensorboard
tensorboard-data-server==0.7.0
tensorboard-plugin-profile==2.13.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.2.1
##git+https://github.com/NVlabs/tiny-cuda-nn/@b916a481d943a9a29d7aaa9519d7024395e98c1c#subdirectory=bindings/torch
##C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==1.13.1
torch-fidelity==0.3.0
torch-tb-profiler==0.4.1
torchmetrics==0.11.3
torchtyping==0.1.4
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.65.0
##traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
typeguard==2.13.3
##typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
tyro==0.4.0
tzdata==2023.3
u-msgpack-python==2.7.2
uri-template==1.2.0
urllib3==1.26.14
wandb==0.13.10
##wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wincertstore==0.2
wrapt==1.14.1
xatlas==0.0.7
y-py==0.5.9
yarg==0.1.9
yarl==1.8.2
ypy-websocket==0.8.2
##zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
